[Silver III] Title: 주유소, Time: 20 ms, Memory: 2804 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -21,8 +21,8 @@ int main()
 	{
 		std::cin >> prices[i];
 	}
-	int curPrice = 1000000000;
-	int sum = 0;
+	long long curPrice = 1000000000;
+	long long sum = 0;
 	for (int i = 0; i < N - 1; ++i)
 	{
 		if (curPrice > prices[i])
